Handle func-based fields and allow expressions in upserts
``pgbulk.upsert`` allows users to provide a ``pgbulk.UpdateField`` object
to the ``update_fields`` argument, allowing a users to specify an expression
that happens if an update occurs.

This allows, for example, a user to do ``models.F('my_field') + 1`` and
increment integer fields in a ``pgbulk.upsert``.

Along with this, fields that cast to ``Func`` and other expressions are
properly handled during upsert.

Type: feature